Move the metadata functions into a separate package
